refactor: Split up transaction modal
