+++ src/components/TransactionConfirmationModal/ConfirmationPendingContent.tsx
@@ -1,0 +1,46 @@
+import React from 'react'
+import styled from 'styled-components'
+import { CloseIcon, Spinner } from '../Shared'
+import { Text } from 'rebass'
+import { RowBetween } from '../Row'
+import { AutoColumn } from '../Column'
+import Circle from '../../assets/images/blue-loader.svg'
+import { Wrapper, Section, ConfirmedIcon } from './helpers'
+
+type ConfirmationPendingContentProps = { onDismiss: () => void; pendingText: string }
+
+const CustomLightSpinner = styled(Spinner)<{ size: string }>`
+  height: ${({ size }) => size};
+  width: ${({ size }) => size};
+`
+
+const ConfirmationPendingContent = ({ onDismiss, pendingText }: ConfirmationPendingContentProps) => {
+  return (
+    <Wrapper>
+      <Section>
+        <RowBetween>
+          <div />
+          <CloseIcon onClick={onDismiss} />
+        </RowBetween>
+        <ConfirmedIcon>
+          <CustomLightSpinner src={Circle} alt="loader" size={'90px'} />
+        </ConfirmedIcon>
+        <AutoColumn gap="12px" justify={'center'}>
+          <Text fontWeight={500} fontSize={20}>
+            Waiting For Confirmation
+          </Text>
+          <AutoColumn gap="12px" justify={'center'}>
+            <Text fontWeight={600} fontSize={14} color="" textAlign="center">
+              {pendingText}
+            </Text>
+          </AutoColumn>
+          <Text fontSize={12} color="#565A69" textAlign="center">
+            Confirm this transaction in your wallet
+          </Text>
+        </AutoColumn>
+      </Section>
+    </Wrapper>
+  )
+}
+
+export default ConfirmationPendingContent
