+++ src/components/TransactionConfirmationModal/helpers.tsx
@@ -1,0 +1,20 @@
+import styled from 'styled-components'
+import { AutoColumn, ColumnCenter } from '../Column'
+
+export const Wrapper = styled.div`
+  width: 100%;
+  overflow-y: auto;
+`
+export const Section = styled(AutoColumn)`
+  padding: 24px;
+`
+
+export const ConfirmedIcon = styled(ColumnCenter)`
+  padding: 60px 0;
+`
+
+export const BottomSection = styled(Section)`
+  background-color: ${({ theme }) => theme.colors.invertedContrast};
+  border-bottom-left-radius: 20px;
+  border-bottom-right-radius: 20px;
+`
